pyproject.toml
[project]
name = "ai-launchpad"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "chainlit>=2.6.8",
    "chromadb>=1.0.20",
    "fastmcp>=2.12.0",
    "langchain-anthropic>=0.3.18",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.5.4",
    "langgraph-sdk>=0.2.0",
    "openai>=1.97.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
    "tavily-python>=0.7.11",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
    "langgraph-cli[inmem]>=0.3.6",
]

[tool.uv]
package = true
